pyproject.toml
[tool.poetry]
name = "oxcompanytechtask"
version = "0.1.0"
description = ""
authors = ["DJmoster <Zapotichnyi12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.112.0"}
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.1.0"
pyjwt = "2.8.0"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
